renamed get_mode.py and used key class
